changed POWERSERIES to work with vectors instead of variadic templates

+++ include/combinatorics.h
@@ -6,7 +6,7 @@
 #define COMBINATORICS_h
 #include <vector>
 namespace iRRAM{
-  std::vector<std::vector<int>> partitions(const int n, const int k);
+  std::vector<std::vector<unsigned long>> partitions(const unsigned long n, const unsigned long k);
   INTEGER factorial(int n);
   REAL choose(int n, int k);
 }
